rename: 패키지 이름 수정 (#53)
패키지 이름이 controller가 아니라 contoller여서 수정했습니다.
