+++ be/src/main/java/codesquad/kr/gyeonggidoidle/issuetracker/domain/issue/controller/request/IssueStatusRequest.java
@@ -1,4 +1,4 @@
-package codesquad.kr.gyeonggidoidle.issuetracker.domain.issue.contoller.request;
+package codesquad.kr.gyeonggidoidle.issuetracker.domain.issue.controller.request;
 
 import codesquad.kr.gyeonggidoidle.issuetracker.domain.issue.service.condition.IssueStatusCondition;
 import com.fasterxml.jackson.annotation.JsonProperty;
